Fix: bean, data type 에러 수정

+++ BOGUdanyo/src/main/java/com/backend3rd/BOGUdanyo/repository/RouteAccidentRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 
 
-public interface RouteAccidentRepository extends JpaRepository<AccidentArea, Long> {
+public interface RouteAccidentRepository extends JpaRepository<AccidentArea, Integer> {
 
     @Query("select a from AccidentArea a where DEGREES(ACOS(SIN(RADIANS(a.lat))*SIN(RADIANS(:middleLat)) + COS(RADIANS(a.lat))*COS(RADIANS(:middleLat))*COS(RADIANS(a.lon-:middleLon))))* 60*1.1515*1609.344 < :radius")
-    List<AccidentArea> findByRegion(@Param(value = "middleLat") double middleLat,
-                                    @Param(value = "middleLon") double middleLon,
+    List<AccidentArea> findByRegion(@Param(value = "middleLat") float middleLat,
+                                    @Param(value = "middleLon") float middleLon,
                                     @Param(value = "radius") double radius);
 }
